refactor(semantics): inline the highlight helpers that had one caller
highlightTarget was one expression at one call site; clearIn was a short block the show that
called it reads better with in place; Rect.isFinite was one condition. The two highlight request
handlers are used by the agent plugin alone, so they are its private members now rather than a
file of their own.
